third_party/blink/web_tests/platform/webview/external/wpt/webhid/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 53 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] HID interface: existence and properties of interface object
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface object length
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface object name
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: attribute onconnect
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: attribute ondisconnect
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: operation getDevices()
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID interface: operation requestDevice(HIDDeviceRequestOptions)
  assert_own_property: self does not have own property "HID" expected property "HID" missing
[FAIL] HID must be primary interface of navigator.hid
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of navigator.hid
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HID interface: navigator.hid must inherit property "onconnect" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HID interface: navigator.hid must inherit property "ondisconnect" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HID interface: navigator.hid must inherit property "getDevices()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HID interface: navigator.hid must inherit property "requestDevice(HIDDeviceRequestOptions)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HID interface: calling requestDevice(HIDDeviceRequestOptions) on navigator.hid with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] HIDDevice interface: existence and properties of interface object
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface object length
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface object name
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute oninputreport
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute opened
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute vendorId
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute productId
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute productName
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: attribute collections
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation open()
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation close()
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation forget()
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation sendReport(octet, BufferSource)
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation sendFeatureReport(octet, BufferSource)
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDDevice interface: operation receiveFeatureReport(octet)
  assert_own_property: self does not have own property "HIDDevice" expected property "HIDDevice" missing
[FAIL] HIDConnectionEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface object length
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface object name
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDConnectionEvent interface: attribute device
  assert_own_property: self does not have own property "HIDConnectionEvent" expected property "HIDConnectionEvent" missing
[FAIL] HIDInputReportEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface object length
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface object name
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: attribute device
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: attribute reportId
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] HIDInputReportEvent interface: attribute data
  assert_own_property: self does not have own property "HIDInputReportEvent" expected property "HIDInputReportEvent" missing
[FAIL] Navigator interface: attribute hid
  assert_true: The prototype object must have a property "hid" expected true got false
[FAIL] Navigator interface: navigator must inherit property "hid" with the proper type
  assert_inherits: property "hid" not found in prototype chain
Harness: the test ran to completion.
